OperatorLearning.FourierLayer — Type

FourierLayer(in, out, grid, modes, σ=identity, init=glorot_uniform) FourierLayer(Wf::AbstractArray, Wl::AbstractArray, [bias_f, bias_l, σ])

Create a Layer of the Fourier Neural Operator as proposed by Li et al. arXiv: 2010.08895

The layer does a fourier transform on the grid dimension of the input array, filters higher modes out by the weight matrix and transforms it to the specified output dimension such that In x M x N -> Out x M x N. The output though only contains the relevant Fourier modes with the rest padded to zero in the last axis as a result of the filtering.

The input x should be a rank 3 tensor of shape (num parameters (in) x num grid points (grid) x batch size (batch)) The output y will be a rank 3 tensor of shape (out x num grid points (grid) x batch size (batch))

You can specify biases for the paths as you like, though the convolutional path is originally not intended to perform an affine transformation.

Examples

Say you're considering a 1D diffusion problem on a 64 point grid. The input is comprised of the grid points as well as the IC at this point. The data consists of 200 instances of the solution. Beforehand we convert the two input channels into a higher-dimensional latent space with 128 nodes by using a regular Dense layer. So the input takes the dimension 128 x 64 x 200. The output would be the diffused variable at a later time, which initially makes the output of the form 128 x 64 x 200 as well. Finally, we have to squeeze this high-dimensional ouptut into the one quantity of interest again by using a Dense layer.

We wish to only keep the first 16 modes of the input and work with the classic sigmoid function as activation.

So we would have:

model = FourierLayer(128, 128, 100, 16, σ)

OperatorLearning.cglorot_normal — Method

cglorotnormal([rng=GLOBALRNG], dims...)

A modification of the glorot_normal function provided by Flux to accommodate Complex numbers. This is necessary since the parameters of the global convolution operator in the Fourier Layer generally has complex weights.

OperatorLearning.cglorot_uniform — Method

cglorotuniform([rng=GLOBALRNG], dims...)

A modification of the glorot_uniform function provided by Flux to accommodate Complex numbers. This is necessary since the parameters of the global convolution operator in the Fourier Layer generally has complex weights.